+++ examples/errors/main.go
@@ -5,8 +5,8 @@ import "github.com/mondegor/go-webcore/mrcore"
 func main() {
 	logger := mrcore.DefaultLogger().DisableFileLine()
 
-	logger.Err(mrcore.FactoryErrInternal.Caller(-1).New())
-	logger.Err(mrcore.FactoryErrInternalTypeAssertion.Caller(-1).New("MY-TYPE", "MY-VALUE"))
+	logger.Err(mrcore.FactoryErrInternal.New())
+	logger.Err(mrcore.FactoryErrInternalTypeAssertion.New("MY-TYPE", "MY-VALUE"))
 
-	logger.Info(mrcore.FactoryErrStorageQueryDataContainer.Caller(-1).New("MY-DATA").Error())
+	logger.Info(mrcore.FactoryErrInternalWithData.New("MY-DATA").Error())
 }