+++ mrcore/errors_http.go
@@ -3,29 +3,14 @@ package mrcore
 import . "github.com/mondegor/go-sysmess/mrerr"
 
 var (
-	FactoryErrHttpRequestParamEmpty = NewFactory(
-		"errHttpRequestParamEmpty", ErrorKindUser, "request param with key '{{ .key }}' is empty'")
-
-	FactoryErrHttpRequestParamMax = NewFactory(
-		"errHttpRequestParamMax", ErrorKindUser, "request param with key '{{ .key }}' has value greater then max '{{ .max }}'")
-
-	FactoryErrHttpRequestParamLenMax = NewFactory(
-		"errHttpRequestParamLenMax", ErrorKindUser, "request param with key '{{ .key }}' has value length greater then max '{{ .maxLength }}'")
-
-	FactoryErrHttpRequestParseParam = NewFactory(
-		"errHttpRequestParseParam", ErrorKindUser, "request param of type '{{ .type }}' with key '{{ .key }}' contains incorrect value '{{ .value }}'")
-
-	FactoryErrHttpRequestParseData = NewFactory(
-		"errHttpRequestParseData", ErrorKindUser, "request body is not valid")
-
 	FactoryErrHttpResponseParseData = NewFactory(
 		"errHttpResponseParseData", ErrorKindInternal, "response data is not valid")
 
 	FactoryErrHttpResponseSendData = NewFactory(
 		"errHttpResponseSendData", ErrorKindInternal, "response data is not send")
 
-	FactoryErrHttpResponseSystemTemporarilyUnableToProcess = NewFactory(
-		"errHttpResponseSystemTemporarilyUnableToProcess", ErrorKindUser, "the system is temporarily unable to process your request. please try again later")
+	FactoryErrHttpMultipartFormFile = NewFactory(
+		"errHttpMultipartFormFile", ErrorKindSystem, "the file with the specified key '{{ .key }}' cannot be processed")
 
 	FactoryErrHttpClientUnauthorized = NewFactory(
 		"errHttpClientUnauthorized", ErrorKindUser, "401. client is unauthorized")
@@ -36,6 +21,18 @@ var (
 	FactoryErrHttpResourceNotFound = NewFactory(
 		"errHttpResourceNotFound", ErrorKindUser, "404. resource not found")
 
-	FactoryErrHttpMultipartFormFile = NewFactory(
-		"errHttpMultipartFormFile", ErrorKindSystem, "the file with the specified key '{{ .key }}' cannot be processed")
+	FactoryErrHttpRequestParseData = NewFactory(
+		"errHttpRequestParseData", ErrorKindUser, "request body is not valid")
+
+	FactoryErrHttpRequestParseParam = NewFactory(
+		"errHttpRequestParseParam", ErrorKindUser, "request param with key '{{ .key }}' of type '{{ .type }}' contains incorrect value '{{ .value }}'")
+
+	FactoryErrHttpRequestParamEmpty = NewFactory(
+		"errHttpRequestParamEmpty", ErrorKindUser, "request param with key '{{ .key }}' is empty'")
+
+	FactoryErrHttpRequestParamMax = NewFactory(
+		"errHttpRequestParamMax", ErrorKindUser, "request param with key '{{ .key }}' contains value greater then max '{{ .max }}'")
+
+	FactoryErrHttpRequestParamLenMax = NewFactory(
+		"errHttpRequestParamLenMax", ErrorKindUser, "request param with key '{{ .key }}' has value length greater then max '{{ .maxLength }}' characters")
 )