+++ mrcore/errors_internal.go
@@ -6,6 +6,9 @@ var (
 	FactoryErrInternal = NewFactory(
 		ErrorInternalID, ErrorKindInternal, "internal server error")
 
+	FactoryErrInternalNotice = NewFactory(
+		ErrorInternalID, ErrorKindInternalNotice, "internal server error")
+
 	FactoryErrInternalNilPointer = NewFactory(
 		"errInternalNilPointer", ErrorKindInternal, "nil pointer")
 
@@ -13,14 +16,14 @@ var (
 		"errInternalTypeAssertion", ErrorKindInternal, "invalid type '{{ .type }}' assertion (value: {{ .value }})")
 
 	FactoryErrInternalInvalidType = NewFactory(
-		"errInternalInvalidType", ErrorKindInternal, "invalid type '{{ .type1 }}', expected: '{{ .type2 }}'")
-
-	FactoryErrInternalInvalidData = NewFactory(
-		"errInternalInvalidData", ErrorKindInternal, "invalid data '{{ .data }}'")
+		"errInternalInvalidType", ErrorKindInternal, "invalid type '{{ .currentType }}', expected: '{{ .expectedType }}'")
 
 	FactoryErrInternalFailedToClose = NewFactory(
 		"errInternalFailedToClose", ErrorKindInternal, "failed to close '{{ .name }}'")
 
-	FactoryErrInternalMapValueNotFound = NewFactory(
-		"errInternalMapValueNotFound", ErrorKindInternal, "'{{ .value }}' is not found in map {{ .name }}")
+	FactoryErrInternalWithData = NewFactory(
+		"errInternalWithData", ErrorKindInternal, "{{ .key }}={{ .data }}")
+
+	FactoryErrWithData = NewFactory(
+		"errWithData", ErrorKindInternalNotice, "{{ .key }}={{ .data }}")
 )