+++ mrreq/int64.go
@@ -30,7 +30,7 @@ func ParseInt64(r *http.Request, key string, required bool) (int64, error) {
 	item, err := strconv.ParseInt(value, 10, 64)
 
 	if err != nil {
-		return 0, mrcore.FactoryErrHttpRequestParseParam.New("Int64", key, value)
+		return 0, mrcore.FactoryErrHttpRequestParseParam.New(key, "Int64", value)
 	}
 
 	return item, nil