+++ mrtool/service_helper.go
@@ -13,7 +13,7 @@ type (
 
 func NewServiceHelper() *ServiceHelper {
 	return &ServiceHelper{
-		callerSkip: 1,
+		callerSkip: 2, // skip: wrapErrorFailed and parent function
 	}
 }
 
@@ -23,44 +23,47 @@ func (h *ServiceHelper) Caller(skip int) *ServiceHelper {
 	}
 }
 
-func (h *ServiceHelper) WrapErrorEntityFetch(err error, entityName string, entityData any) error {
-	if mrcore.FactoryErrStorageNoRowFound.Is(err) {
-		return mrcore.FactoryErrServiceEntityNotFound.Wrap(err, mrerr.Arg{entityName: entityData})
-	}
-
-	return mrcore.FactoryErrServiceTemporarilyUnavailable.Caller(h.callerSkip).Wrap(err, mrerr.Arg{entityName: entityData})
+func (h *ServiceHelper) IsNotFound(err error) bool {
+	return mrcore.FactoryErrStorageNoRowFound.Is(err) ||
+		mrcore.FactoryErrStorageRowsNotAffected.Is(err)
 }
 
-func (h *ServiceHelper) WrapErrorEntityInsert(err error, entityName string, entityData any) error {
-	if mrcore.FactoryErrStorageQueryFailed.Is(err) {
-		return mrcore.FactoryErrServiceEntityNotCreated.Wrap(err, mrerr.Arg{entityName: entityData})
-	}
-
-	return mrcore.FactoryErrServiceTemporarilyUnavailable.Caller(h.callerSkip).Wrap(err, mrerr.Arg{entityName: entityData})
+func (h *ServiceHelper) WrapErrorFailed(err error, source string) error {
+	return h.wrapErrorFailed(err, "source", source)
 }
 
-func (h *ServiceHelper) WrapErrorEntityUpdate(err error, entityName string, entityData any) error {
-	if mrcore.FactoryErrStorageRowsNotAffected.Is(err) ||
-		mrcore.FactoryErrStorageNoRowFound.Is(err) {
-		return mrcore.FactoryErrServiceEntityNotFound.Wrap(err, mrerr.Arg{entityName: entityData})
-	}
-
-	if mrcore.FactoryErrStorageQueryFailed.Is(err) {
-		return mrcore.FactoryErrServiceEntityNotStored.Wrap(err, mrerr.Arg{entityName: entityData})
-	}
-
-	return mrcore.FactoryErrServiceTemporarilyUnavailable.Caller(h.callerSkip).Wrap(err, mrerr.Arg{entityName: entityData})
+func (h *ServiceHelper) WrapErrorEntityFailed(err error, entityName string, entityData any) error {
+	return h.wrapErrorFailed(err, entityName, entityData)
 }
 
-func (h *ServiceHelper) WrapErrorEntityDelete(err error, entityName string, entityData any) error {
-	if mrcore.FactoryErrStorageRowsNotAffected.Is(err) ||
-		mrcore.FactoryErrStorageNoRowFound.Is(err) {
-		return mrcore.FactoryErrServiceEntityNotFound.Wrap(err, mrerr.Arg{entityName: entityData})
+func (h *ServiceHelper) WrapErrorEntity(
+	reasonWhyNotFound *mrerr.AppErrorFactory,
+	err error,
+	entityName string,
+	entityData any,
+) error {
+	if h.IsNotFound(err) {
+		return reasonWhyNotFound.Wrap(err)
 	}
+
+	return h.wrapErrorFailed(err, entityName, entityData)
+}
+
+func (h *ServiceHelper) WrapErrorEntityNotFoundOrFailed(err error, entityName string, entityData any) error {
+	return h.WrapErrorEntity(
+		mrcore.FactoryErrServiceEntityNotFound,
+		err,
+		entityName,
+		entityData,
+	)
+}
+
+func (h *ServiceHelper) wrapErrorFailed(err error, name string, data any) error {
+	err = mrcore.FactoryErrWithData.Wrap(err, name, data)
 
 	if mrcore.FactoryErrStorageQueryFailed.Is(err) {
-		return mrcore.FactoryErrServiceEntityNotRemoved.Wrap(err, mrerr.Arg{entityName: entityData})
+		return mrcore.FactoryErrServiceOperationFailed.Wrap(err)
 	}
 
-	return mrcore.FactoryErrServiceTemporarilyUnavailable.Caller(h.callerSkip).Wrap(err, mrerr.Arg{entityName: entityData})
+	return mrcore.FactoryErrServiceTemporarilyUnavailable.Caller(h.callerSkip).Wrap(err)
 }